requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.12
contextlib2==21.6.0
docopt==0.6.2
idna==3.3
internetarchive==3.0.0
jsonpatch==1.32
jsonpointer==2.3
requests==2.27.1
schema==0.7.5
tqdm==4.64.0
urllib3==1.26.9
pyyaml
pydantic
